Add various example lua scripts and enhance editor UI
- Updated CMakeLists.txt to support 64-bit builds with MSVC.
- Fixed line counting in ScriptComponent to ensure correct return type.
- Enhanced EditorUI with playback controls for better scene management.
- Improved Window class to handle resizing and title updates dynamically.
- Refactored main.cpp to include new script components and handle window resize events.

+++ src/Components/ScriptComponent.cpp
@@ -455,7 +455,7 @@ namespace spark
 
     int ScriptComponent::CountLines() const
     {
-        return std::count(m_scriptContent.begin(), m_scriptContent.end(), '\n') + 1;
+        return static_cast<int>(std::count(m_scriptContent.begin(), m_scriptContent.end(), '\n') + 1);
     }
 
     bool ScriptComponent::ReloadScript()
